av1-etl-spotify-youtube/requirements.txt
spotipy==2.22.1
google-api-python-client==2.77.0
pandas==2.2.3
sqlalchemy==2.1.4
streamlit==1.26.0
python-dotenv==1.0.0
requests==2.31.0

psycopg2-binary==2.9.7
matplotlib==3.8.1